requirements.txt
Django==1.10.3
djangorestframework==3.5.3
httpie==0.9.6
Pygments==2.1.3
requests==2.12.1
